+++ ProjectJAVA/src/main/java/com/example/ProjectJAVA/Service/Imp/ProductServiceImp.java
@@ -4,6 +4,7 @@ import com.example.ProjectJAVA.DTO.ProductDTO;
 import com.example.ProjectJAVA.Entity.Products;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ProductServiceImp {
@@ -12,7 +13,7 @@ public interface ProductServiceImp {
 
     Products getProductById(int product_id);
 
-    Boolean createProduct(MultipartFile file, String product_name, String description, int quantity, int price, int category_id);
+    Boolean createProduct(MultipartFile file, String product_name, String description, int quantity, BigDecimal price, int category_id);
 
 
 }
